PythonApp/requirements.txt
pyautogui
flask
qrcode==6.1
Pillow
qrcode-terminal